requirements.txt
dask[complete]
ray[tune]
numpy
scipy
scikit-learn
pandas
statsmodels
catboost
pymysql
sqlalchemy
scikit-optimize~=0.9.0
opencv-python~=4.7.0.68
torch==2.0.1
pytorch-lightning~=2.0.4
open_clip_torch~=2.20.0
pillow~=9.4.0
torchvision~=0.15.2
transformers
imutils
joblib
shap
tqdm
numba
requests
boto3
pydantic
python-json-logger
aiohttp
fastapi
s3fs
